Created methods and routes for game logic, closes #62, closes #63
Created GET /games/:id/stream for sending data to the user. It listens for events with game+gameid (game41), and then sends the user the field that was changed.
Created POST /games/:id for collecting data. It checks the setp for validity and prepends (adds to the begining) the step to game.step arrray (which is in memory).
Modified POST / for compatibility
Created middleware/validStep.js fir step validation